+++ src/Resource/IdentificationInfo.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCz\DigiSign\Resource;
+
+use DateTime;
+
+class IdentificationInfo extends BaseResource
+{
+    public string $id;
+    public string $status;
+    public ?DateTime $completedAt;
+}